hackernews useConfig
